+++ pom.xml
@@ -2,16 +2,11 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>dubbo-registry</artifactId>
-        <groupId>com.alibaba</groupId>
-        <version>2.6.5</version>
-    </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.alibaba</groupId>
+	<groupId>org.apache.dubbo</groupId>
     <artifactId>dubbo-registry-nacos</artifactId>
-    <version>0.0.2</version>
+    <version>0.0.3</version>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -19,21 +14,21 @@
         <!-- Compiler settings properties -->
         <maven.compiler.source>1.6</maven.compiler.source>
         <maven.compiler.target>1.6</maven.compiler.target>
-        <dubbo.version>2.6.5</dubbo.version>
-        <nacos.version>0.6.1</nacos.version>
+        <dubbo.version>2.7.0</dubbo.version>
+        <nacos.version>0.7.0</nacos.version>
     </properties>
 
     <dependencies>
 
         <dependency>
-            <groupId>com.alibaba</groupId>
+            <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-registry-api</artifactId>
             <version>${dubbo.version}</version>
             <optional>true</optional>
         </dependency>
 
         <dependency>
-            <groupId>com.alibaba</groupId>
+            <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-common</artifactId>
             <version>${dubbo.version}</version>
             <optional>true</optional>
@@ -48,35 +43,35 @@
 
         <!-- Test Libraries -->
         <dependency>
-            <groupId>com.alibaba</groupId>
+            <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-config-api</artifactId>
             <version>${dubbo.version}</version>
             <scope>test</scope>
         </dependency>
 
         <dependency>
-            <groupId>com.alibaba</groupId>
+            <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-serialization-hessian2</artifactId>
             <version>${dubbo.version}</version>
             <scope>test</scope>
         </dependency>
 
         <dependency>
-            <groupId>com.alibaba</groupId>
+            <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-config-spring</artifactId>
             <version>${dubbo.version}</version>
             <scope>test</scope>
         </dependency>
 
         <dependency>
-            <groupId>com.alibaba</groupId>
+            <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-rpc-dubbo</artifactId>
             <version>${dubbo.version}</version>
             <scope>test</scope>
         </dependency>
 
         <dependency>
-            <groupId>com.alibaba</groupId>
+            <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo-remoting-netty</artifactId>
             <version>${dubbo.version}</version>
             <scope>test</scope>
@@ -92,6 +87,7 @@
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-test</artifactId>
+            <version>5.1.4.RELEASE</version>
             <scope>test</scope>
         </dependency>
