+++ src/test/java/org/apache/dubbo/demo/provider/DemoServiceProviderXmlBootstrap.java
@@ -14,12 +14,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.registry.nacos;
+package org.apache.dubbo.demo.provider;
+
+import org.apache.dubbo.demo.service.DemoService;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
+
+import java.io.IOException;
 
 /**
- * {@link NacosRegistryFactory} Test
- *
- * @since 2.6.5
+ * {@link DemoService} provider demo XML bootstrap
  */
-public class NacosRegistryFactoryTest {
+public class DemoServiceProviderXmlBootstrap {
+
+    public static void main(String[] args) throws IOException {
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext();
+        context.setConfigLocation("/META-INF/spring/dubbo-provider-context.xml");
+        context.refresh();
+        System.out.println("DemoService provider (XML) is starting...");
+        System.in.read();
+    }
 }
